Updated .gitignore and re-added all files into the index

+++ app.js
@@ -154,35 +154,49 @@ http.createServer(app).listen(app.get('port'), function(){
 var fs = require('fs')
 	util = require('util');
 
-setInterval(function() {
-	fs.readdir(app.get('tempDir'), function(err, files) {
+// Create temp directory
+fs.mkdir(app.get('tempDir'), function(err) {
 
-		if (err) {
-			winston.error("Couldn't read files from the temporary directory: %s", app.get('tempDir'));
-		} else {
+	if (err && err.code != 'EEXIST') {
 
-			/*winston.info("Following files were found in the temporary directory: %s", app.get('tempDir'));
-			console.log(util.inspect(files));*/
+		winston.error("We couldn't create the temporary directory: %s", app.get('tempDir'));
+		console.log(util.inspect(err, {colors: true}));
 
-			files.forEach(function(file) {
-				fs.stat(app.get('tempDir') + '/' + file, function(err, stats) {
+	} else {
 
-					if (err) { 
-						winston.error("Couldn't read stats for file: %s", app.get('tempDir') + '/' + file); 
-						return;
-					}
+		winston.info("Created temporary directory: %s", app.get('tempDir'));
 
-					if (stats.mtime < (new Date(Date.now() - 1000 * 60 * 5)).getTime()) {
-						fs.unlink(app.get('tempDir') + '/' + file, function(err) {
+		setInterval(function() {
+			fs.readdir(app.get('tempDir'), function(err, files) {
 
-							if (err) { winston.error("Couldn't delete file: %s", app.get('tempDir') + '/' + file); }
-							else {
-								winston.info("Deleted file: %s", app.get('tempDir') + '/' + file);
+				if (err) {
+					winston.error("Couldn't read files from the temporary directory: %s", app.get('tempDir'));
+				} else {
+
+					/*winston.info("Following files were found in the temporary directory: %s", app.get('tempDir'));
+					console.log(util.inspect(files));*/
+
+					files.forEach(function(file) {
+						fs.stat(app.get('tempDir') + '/' + file, function(err, stats) {
+
+							if (err) { 
+								winston.error("Couldn't read stats for file: %s", app.get('tempDir') + '/' + file); 
+								return;
+							}
+
+							if (stats.mtime < (new Date(Date.now() - 1000 * 60 * 5)).getTime()) {
+								fs.unlink(app.get('tempDir') + '/' + file, function(err) {
+
+									if (err) { winston.error("Couldn't delete file: %s", app.get('tempDir') + '/' + file); }
+									else {
+										winston.info("Deleted file: %s", app.get('tempDir') + '/' + file);
+									}
+								});
 							}
 						});
-					}
-				});
+					});
+				}
 			});
-		}
-	});
-}, 10000);
+		}, 10000);
+	}
+});
